Revise some enums and refactor Pipeline class
Revised `BlendFactor`, `BlendOperation`, and other enumerations by adding a string value to each entry for the sake of easier readability and understanding. The `Pipeline` class underwent substantial refactoring, which included setting default values for member variables and changing types of some variables from nullable to non-nullable. Furthermore, the `GPUSampleMask` and `GPUStencilValue` type aliases are switched from `Int` to `Long` to accommodate larger data sizes. This commit aims to improve code readability and maintainability.

+++ wgpu4k/src/commonMain/kotlin/io.ygdrasil.wgpu/Types.kt
@@ -24,7 +24,7 @@ typealias GPUMapModeFlags = Number
 
 typealias GPUPipelineConstantValue = Number
 
-typealias GPUSampleMask = Int
+typealias GPUSampleMask = Long
 
 typealias GPUShaderStageFlags = Number
 
@@ -38,7 +38,7 @@ typealias GPUSize64 = Long
 
 typealias GPUSize64Out = Long
 
-typealias GPUStencilValue = Int
+typealias GPUStencilValue = Long
 
 typealias GPUTextureUsageFlags = Int
